backend/requirements.txt
# Core FastAPI Framework
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Configuration & Environment
python-dotenv>=1.0.0
pydantic>=2.5.0,<3.0.0
pydantic-settings>=2.0.0

# HTTP Clients
httpx>=0.25.0
aiohttp>=3.9.0

# Document Processing
pypdf>=3.17.0
python-docx>=1.1.0

# Translation
# Use deep-translator (more maintained) instead of googletrans
deep-translator>=1.11.0
langdetect>=1.0.9

# Deep Learning Framework (Install before speech models)
torch>=2.1.0,<2.3.0
torchaudio>=2.1.0,<2.3.0

# Speech Processing
openai-whisper>=20231117
soundfile>=0.12.1
numpy<2.0.0

# Parler TTS (Install after torch)
git+https://github.com/huggingface/parler-tts.git

# Transformers & NLP
transformers>=4.35.0
accelerate>=0.25.0
sentencepiece>=0.1.99

# Video/Audio Processing
ffmpeg-python>=0.2.0
pydub>=0.25.1

# Database
sqlalchemy>=2.0.23
aiosqlite>=0.19.0

# Additional Utilities
Pillow>=10.0.0
regex>=2023.10.3

